Reafactoring and renamings
